Code for App global state was created.
The global state is being managed using the store pattern,
an event subscription is made, an event or events handlers are called when the
todo list data is changed.

AppStore.js located in app/store folder contains the store which controls the
application state.

PubSub.js located in lib contains the code to create and subscribe to an event
also contains the code to publish an event. when the event is published
the handlers subscribe to that event are called and receives the data of the
todo list updated.

actions.js located in app/modules/task-management folder contains the code to
request operations to the store that update the global state (the todo-list data),
the operations that the code of "actions.js can request are,
add an item,delete an item  and mark an item as finished, the code in this file.

the store finally in the commit function interacts with "mutations.js" code,
the functions in this file finally make the changes to the application state.

The function called "dispatch" located in the store, is the function which will be
used in UI components to make changes to the application state, receive as
arguments an action key, which is
the action that need to be made (add a todo,remove a one or update it)
and a piece of data which will be used to make the change of the state.

state.js located in app/store/ folder is the file which contain the data (app state)
comprises an array which have the todo items.